billing-api/requirements.txt
fastapi==0.95.1
orjson==3.7.7
pydantic==1.10.7
pydantic[email]==1.10.7
uvicorn==0.21.1
uvloop==0.17.0
flake8==4.0.1
gunicorn==20.1.0
SQLAlchemy==2.0.0
psycopg2-binary==2.9.3
sqlalchemy[asyncio]==2.0.0
asyncpg==0.27.0 
python-jose[cryptography]==3.3.0
requests==2.28.1
oitnb==0.2.2
celery==5.2.7
stripe==5.3.0
async-stripe==5.3.0
grpcio==1.53.0
grpcio-tools==1.53.0
asgiref==3.6.0
backoff==2.2.1 
